requirements.txt
discord.py
discord.py[voice]
youtube-dl
google-api-python-client
asyncio